Fix: Extract SOAP URL from wsdl URL passed
Extract SOAP URL from wsdl URL because the URL present in WSDL might not be the same and many times we dont have control over it. It might happen that the WSDL URL could be https but in wsdl definition, it could start with http. Calling https URL with http will give error below:

Expected element type <Envelope> but have <html>